refactor: DIP를 위한 추상화 리팩터링
