+++ Api/src/main/java/com/dingdong/api/image/controller/ImageController.java
@@ -1,7 +1,6 @@
 package com.dingdong.api.image.controller;
 
 
-import com.dingdong.api.image.controller.request.DeleteImageRequest;
 import com.dingdong.api.image.controller.response.UploadImageResponse;
 import com.dingdong.api.image.service.ImageService;
 import io.swagger.v3.oas.annotations.Operation;
@@ -11,23 +10,16 @@ import org.springframework.http.MediaType;
 import org.springframework.web.bind.annotation.*;
 import org.springframework.web.multipart.MultipartFile;
 
-@Tag(name = "이미지 업로드")
+@Tag(name = "이미지")
 @RestController
-@RequestMapping("/upload")
+@RequestMapping("/images")
 @RequiredArgsConstructor
 public class ImageController {
     private final ImageService imageService;
 
-    @Operation(summary = "이미지 업로드")
-    @PostMapping(path = "/image", consumes = MediaType.MULTIPART_FORM_DATA_VALUE)
-    public UploadImageResponse uploadImage(@RequestParam("image") MultipartFile multipartFile)
-            throws Exception {
+    @Operation(summary = "업로드")
+    @PostMapping(consumes = MediaType.MULTIPART_FORM_DATA_VALUE)
+    public UploadImageResponse uploadImage(@RequestParam("image") MultipartFile multipartFile) {
         return UploadImageResponse.from(imageService.uploadImage(multipartFile));
-    }
-
-    @Operation(summary = "이미지 삭제")
-    @DeleteMapping("/image")
-    public void remove(@RequestBody DeleteImageRequest request) throws Exception {
-        imageService.archive(request.getImageUrl());
     }
 }
