+++ Domain/src/main/java/com/dingdong/domain/domains/image/adaptor/ImageAdaptor.java
@@ -2,7 +2,7 @@ package com.dingdong.domain.domains.image.adaptor;
 
 
 import com.dingdong.core.annotation.Adaptor;
-import com.dingdong.domain.domains.image.domain.Image;
+import com.dingdong.domain.domains.image.domain.DeleteImage;
 import com.dingdong.domain.domains.image.repository.ImageRepository;
 import lombok.RequiredArgsConstructor;
 
@@ -11,7 +11,7 @@ import lombok.RequiredArgsConstructor;
 public class ImageAdaptor {
     private final ImageRepository imageRepository;
 
-    public void save(Image image) {
-        imageRepository.save(image);
+    public void save(DeleteImage deleteImage) {
+        imageRepository.save(deleteImage);
     }
 }
