+++ Domain/src/main/java/com/dingdong/domain/domains/image/repository/ImageRepository.java
@@ -1,9 +1,9 @@
 package com.dingdong.domain.domains.image.repository;
 
 
-import com.dingdong.domain.domains.image.domain.Image;
+import com.dingdong.domain.domains.image.domain.DeleteImage;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface ImageRepository extends JpaRepository<Image, Long> {}
+public interface ImageRepository extends JpaRepository<DeleteImage, Long> {}
