+++ Infrastructure/src/main/java/com/dingdong/infrastructure/image/ImageHandler.java
@@ -1,0 +1,10 @@
+package com.dingdong.infrastructure.image;
+
+
+import org.springframework.web.multipart.MultipartFile;
+
+public interface ImageHandler {
+    String uploadImage(MultipartFile multipartFile);
+
+    void removeImage(String imageUrl);
+}
